refactor(index): change page title's HTML tag to h1
First heading on site index was wrapped in a set of `<h2></h2>` tags`.
I believe that the reason that this element had been placed within the
`<h2>` tags was because the prior author desired the look of an `<h2>`
element. But HTML is not the correct tool for styling elements. HTML
should only be used to signify the element's purpose in the document.
So, I replaced the `<h2>` tag with appropriate `<h1>` tag and added a
`<span class='h2'>` tag to give the element the desired styling of the
h2 element.