build(container): create container configuration
Created file to install a remote Kubernetes cluster from inside a dev
container. Includes kubectl, Helm, minikube, and the Docker.

Alternatively, you can directly build and run the app inside the dev
container you create as opposed to deploying a container image into a
local minikube or remote Kubernetes cluster. However, the latter has
the advantage of not affecting your application and is great for tests.

Signed-off-by: <contact@bethanyfannin.com>