refactor(css): change bg color to a css var
I realized the body's background was a hardcoded hex code. I decided
it would make more sense to make this a variable as well, so I moved it
up with the other CSS variables at the top of style.css, renamed the
former --base variable to --highlight and named the #193549 hex code
--base, since it seemed appropriate for the background color to have
--base as a variable name.

This way, if I use the color again throughout the site, it will be much
simpler to change the color if I decide to in the future.