style(array.css): adjust css variable color names
Change the names of the CSS color variable names to names that
describe the shade of the color to make them more semantic and easy
to work with.

If a variable like "--highlight" is necessary in the future,simply
assign it the color variable. That way if that particular feature ever
changes, it only has to be changed once. Additionally, using color
variable names that describe tone is easier to remember.

Signed-off-by: Bethany <contact@bethanyfannin.com>