refactor: create proper file structure for node
Remove files from root directory and create a 'src' directory for all
files that will be handled by node on the frontend and a 'dist'
directory for the files that will be created. 'Dist' directory is not
tracked with the project.

The 'src' directory is further organized by a 'scripts' directory and
a 'styles' directory. Because there will be multiple stylesheets and
scripts, these directories will make the project easier to navigate
and work with.

Signed-off-by: Bethany Fannin <contact@bethanyfannin.com>