refactor: configure Array exercise to Node.js
Move the files associated with the array-cardio.html page into the
src/ directory, placing the CSS file within src/styles and the JS
file within src/scripts.

Moved the array-cardio.html file to dist/ which is not being tracked.
Kept this HTML file for reference at this point, may use it or use a
framework like React or Next.js.

Signed-off-by: Bethany Fannin <contact@bethanyfannin.com>